+++ grakn-test/src/test/java/ai/grakn/test/graql/query/MatchQueryModifierTest.java
@@ -34,9 +34,9 @@ import static ai.grakn.graql.Graql.or;
 import static ai.grakn.graql.Graql.var;
 import static ai.grakn.graql.Order.asc;
 import static ai.grakn.graql.Order.desc;
-import static ai.grakn.test.matcher.QueryUtil.*;
-import static ai.grakn.test.matcher.QueryUtil.containsAllMovies;
-import static ai.grakn.test.matcher.QueryUtil.variable;
+import static ai.grakn.test.matcher.GraknMatchers.*;
+import static ai.grakn.test.matcher.GraknMatchers.containsAllMovies;
+import static ai.grakn.test.matcher.GraknMatchers.variable;
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.junit.Assert.assertEquals;
