+++ grakn-test/src/test/java/ai/grakn/test/graql/query/MatchQueryTest.java
@@ -65,68 +65,68 @@ import static ai.grakn.graql.Graql.neq;
 import static ai.grakn.graql.Graql.or;
 import static ai.grakn.graql.Graql.regex;
 import static ai.grakn.graql.Graql.var;
-import static ai.grakn.test.matcher.QueryUtil.aRuleType;
-import static ai.grakn.test.matcher.QueryUtil.action;
-import static ai.grakn.test.matcher.QueryUtil.alPacino;
-import static ai.grakn.test.matcher.QueryUtil.allVariables;
-import static ai.grakn.test.matcher.QueryUtil.apocalypseNow;
-import static ai.grakn.test.matcher.QueryUtil.benjaminLWillard;
-import static ai.grakn.test.matcher.QueryUtil.betteMidler;
-import static ai.grakn.test.matcher.QueryUtil.character;
-import static ai.grakn.test.matcher.QueryUtil.chineseCoffee;
-import static ai.grakn.test.matcher.QueryUtil.cluster;
-import static ai.grakn.test.matcher.QueryUtil.comedy;
-import static ai.grakn.test.matcher.QueryUtil.concept;
-import static ai.grakn.test.matcher.QueryUtil.constraintRule;
-import static ai.grakn.test.matcher.QueryUtil.containsAllMovies;
-import static ai.grakn.test.matcher.QueryUtil.crime;
-import static ai.grakn.test.matcher.QueryUtil.drama;
-import static ai.grakn.test.matcher.QueryUtil.entity;
-import static ai.grakn.test.matcher.QueryUtil.family;
-import static ai.grakn.test.matcher.QueryUtil.fantasy;
-import static ai.grakn.test.matcher.QueryUtil.gender;
-import static ai.grakn.test.matcher.QueryUtil.genre;
-import static ai.grakn.test.matcher.QueryUtil.genreOfProduction;
-import static ai.grakn.test.matcher.QueryUtil.godfather;
-import static ai.grakn.test.matcher.QueryUtil.harry;
-import static ai.grakn.test.matcher.QueryUtil.hasTitle;
-import static ai.grakn.test.matcher.QueryUtil.hasType;
-import static ai.grakn.test.matcher.QueryUtil.hasValue;
-import static ai.grakn.test.matcher.QueryUtil.heat;
-import static ai.grakn.test.matcher.QueryUtil.hocusPocus;
-import static ai.grakn.test.matcher.QueryUtil.inferenceRule;
-import static ai.grakn.test.matcher.QueryUtil.isCasting;
-import static ai.grakn.test.matcher.QueryUtil.isInstance;
-import static ai.grakn.test.matcher.QueryUtil.judeLaw;
-import static ai.grakn.test.matcher.QueryUtil.kermitTheFrog;
-import static ai.grakn.test.matcher.QueryUtil.language;
-import static ai.grakn.test.matcher.QueryUtil.marlonBrando;
-import static ai.grakn.test.matcher.QueryUtil.martinSheen;
-import static ai.grakn.test.matcher.QueryUtil.mirandaHeart;
-import static ai.grakn.test.matcher.QueryUtil.missPiggy;
-import static ai.grakn.test.matcher.QueryUtil.movie;
-import static ai.grakn.test.matcher.QueryUtil.movies;
-import static ai.grakn.test.matcher.QueryUtil.musical;
-import static ai.grakn.test.matcher.QueryUtil.name;
-import static ai.grakn.test.matcher.QueryUtil.neilMcCauley;
-import static ai.grakn.test.matcher.QueryUtil.person;
-import static ai.grakn.test.matcher.QueryUtil.production;
-import static ai.grakn.test.matcher.QueryUtil.realName;
-import static ai.grakn.test.matcher.QueryUtil.releaseDate;
-import static ai.grakn.test.matcher.QueryUtil.resource;
-import static ai.grakn.test.matcher.QueryUtil.results;
-import static ai.grakn.test.matcher.QueryUtil.robertDeNiro;
-import static ai.grakn.test.matcher.QueryUtil.rule;
-import static ai.grakn.test.matcher.QueryUtil.runtime;
-import static ai.grakn.test.matcher.QueryUtil.sarah;
-import static ai.grakn.test.matcher.QueryUtil.sarahJessicaParker;
-import static ai.grakn.test.matcher.QueryUtil.spy;
-import static ai.grakn.test.matcher.QueryUtil.theMuppets;
-import static ai.grakn.test.matcher.QueryUtil.title;
-import static ai.grakn.test.matcher.QueryUtil.tmdbVoteAverage;
-import static ai.grakn.test.matcher.QueryUtil.tmdbVoteCount;
-import static ai.grakn.test.matcher.QueryUtil.variable;
-import static ai.grakn.test.matcher.QueryUtil.war;
+import static ai.grakn.test.matcher.GraknMatchers.aRuleType;
+import static ai.grakn.test.matcher.GraknMatchers.action;
+import static ai.grakn.test.matcher.GraknMatchers.alPacino;
+import static ai.grakn.test.matcher.GraknMatchers.allVariables;
+import static ai.grakn.test.matcher.GraknMatchers.apocalypseNow;
+import static ai.grakn.test.matcher.GraknMatchers.benjaminLWillard;
+import static ai.grakn.test.matcher.GraknMatchers.betteMidler;
+import static ai.grakn.test.matcher.GraknMatchers.character;
+import static ai.grakn.test.matcher.GraknMatchers.chineseCoffee;
+import static ai.grakn.test.matcher.GraknMatchers.cluster;
+import static ai.grakn.test.matcher.GraknMatchers.comedy;
+import static ai.grakn.test.matcher.GraknMatchers.concept;
+import static ai.grakn.test.matcher.GraknMatchers.constraintRule;
+import static ai.grakn.test.matcher.GraknMatchers.containsAllMovies;
+import static ai.grakn.test.matcher.GraknMatchers.crime;
+import static ai.grakn.test.matcher.GraknMatchers.drama;
+import static ai.grakn.test.matcher.GraknMatchers.entity;
+import static ai.grakn.test.matcher.GraknMatchers.family;
+import static ai.grakn.test.matcher.GraknMatchers.fantasy;
+import static ai.grakn.test.matcher.GraknMatchers.gender;
+import static ai.grakn.test.matcher.GraknMatchers.genre;
+import static ai.grakn.test.matcher.GraknMatchers.genreOfProduction;
+import static ai.grakn.test.matcher.GraknMatchers.godfather;
+import static ai.grakn.test.matcher.GraknMatchers.harry;
+import static ai.grakn.test.matcher.GraknMatchers.hasTitle;
+import static ai.grakn.test.matcher.GraknMatchers.hasType;
+import static ai.grakn.test.matcher.GraknMatchers.hasValue;
+import static ai.grakn.test.matcher.GraknMatchers.heat;
+import static ai.grakn.test.matcher.GraknMatchers.hocusPocus;
+import static ai.grakn.test.matcher.GraknMatchers.inferenceRule;
+import static ai.grakn.test.matcher.GraknMatchers.isCasting;
+import static ai.grakn.test.matcher.GraknMatchers.isInstance;
+import static ai.grakn.test.matcher.GraknMatchers.judeLaw;
+import static ai.grakn.test.matcher.GraknMatchers.kermitTheFrog;
+import static ai.grakn.test.matcher.GraknMatchers.language;
+import static ai.grakn.test.matcher.GraknMatchers.marlonBrando;
+import static ai.grakn.test.matcher.GraknMatchers.martinSheen;
+import static ai.grakn.test.matcher.GraknMatchers.mirandaHeart;
+import static ai.grakn.test.matcher.GraknMatchers.missPiggy;
+import static ai.grakn.test.matcher.GraknMatchers.movie;
+import static ai.grakn.test.matcher.GraknMatchers.movies;
+import static ai.grakn.test.matcher.GraknMatchers.musical;
+import static ai.grakn.test.matcher.GraknMatchers.name;
+import static ai.grakn.test.matcher.GraknMatchers.neilMcCauley;
+import static ai.grakn.test.matcher.GraknMatchers.person;
+import static ai.grakn.test.matcher.GraknMatchers.production;
+import static ai.grakn.test.matcher.GraknMatchers.realName;
+import static ai.grakn.test.matcher.GraknMatchers.releaseDate;
+import static ai.grakn.test.matcher.GraknMatchers.resource;
+import static ai.grakn.test.matcher.GraknMatchers.results;
+import static ai.grakn.test.matcher.GraknMatchers.robertDeNiro;
+import static ai.grakn.test.matcher.GraknMatchers.rule;
+import static ai.grakn.test.matcher.GraknMatchers.runtime;
+import static ai.grakn.test.matcher.GraknMatchers.sarah;
+import static ai.grakn.test.matcher.GraknMatchers.sarahJessicaParker;
+import static ai.grakn.test.matcher.GraknMatchers.spy;
+import static ai.grakn.test.matcher.GraknMatchers.theMuppets;
+import static ai.grakn.test.matcher.GraknMatchers.title;
+import static ai.grakn.test.matcher.GraknMatchers.tmdbVoteAverage;
+import static ai.grakn.test.matcher.GraknMatchers.tmdbVoteCount;
+import static ai.grakn.test.matcher.GraknMatchers.variable;
+import static ai.grakn.test.matcher.GraknMatchers.war;
 import static ai.grakn.util.ErrorMessage.MATCH_INVALID;
 import static ai.grakn.util.Schema.MetaSchema.RULE;
 import static java.util.stream.Collectors.toList;
