+++ grakn-test/src/test/java/ai/grakn/test/graql/query/QueryBuilderTest.java
@@ -34,8 +34,8 @@ import org.junit.rules.ExpectedException;
 import static ai.grakn.graql.Graql.insert;
 import static ai.grakn.graql.Graql.match;
 import static ai.grakn.graql.Graql.var;
-import static ai.grakn.test.matcher.QueryUtil.containsAllMovies;
-import static ai.grakn.test.matcher.QueryUtil.variable;
+import static ai.grakn.test.matcher.GraknMatchers.containsAllMovies;
+import static ai.grakn.test.matcher.GraknMatchers.variable;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
